% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_theme_base.R
\name{ssz_theme_base}
\alias{ssz_theme_base}
\title{ssz_theme_base}
\usage{
ssz_theme_base(base_size, base_family, base_line_size, base_rect_size)
}
\arguments{
\item{base_size}{optional, basic font size, parameter passed to minimal theme}

\item{base_family}{optional, basic font family, parameter to be passed to minimal theme}

\item{base_line_size}{optional, basic line size, parameter passed to minimal theme}

\item{base_rect_size}{optional, basic rect size, parameter passed to minimal theme}
}
\value{
a ggplot theme
}
\description{
this creates a ggplot theme that can be used as a base theme on top of which to
replace the items that depend on the orientation of the plot
}
