% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_theme_web_base.R
\name{ssz_theme_web_base}
\alias{ssz_theme_web_base}
\title{ssz_theme_base_web}
\usage{
ssz_theme_web_base(base_size, base_family, base_line_size, base_rect_size)
}
\arguments{
\item{base_size}{optional, defines the base font size used throughout the plot.
This parameter is passed to the theme_bw function, ensuring consistent typography
across ggplot elements.}

\item{base_family}{optional, specifies the base font family for the text elements
in the ggplot object. This parameter is passed to the theme_bw function, ensuring
a consistent font style across ggplot elements.}

\item{base_line_size}{optional, sets the base line size for lines and borders
within the ggplot object. This parameter is passed to the theme_bw function, ensuring
consistent line thickness across visual elements.}

\item{base_rect_size}{optional, specifies the base size for rectangular elements
such as legend boxes or plot backgrounds within the ggplot object. This parameter
is passed to the theme_bw function, ensuring consistent sizing of rectangular
shapes across ggplot elements.}
}
\value{
a ggplot theme
}
\description{
Function that creates a ggplot theme that can be used as a base
theme on top of which to replace the theme's items that depend on the orientation
of the web plot.
}
