% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_theme.R
\name{ssz_theme}
\alias{ssz_theme}
\title{ssz_theme}
\usage{
ssz_theme(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  ausrichtung = "x"
)
}
\arguments{
\item{base_size}{optional, basic font size, base_size parameter passed to minimal theme}

\item{base_family}{optional, base_family parameter to be passed to minimal theme}

\item{base_line_size}{tbd}

\item{base_rect_size}{tbd}

\item{ausrichtung}{tbd}
}
\description{
Function for styling ggplots according to Stadt Zürich's corporate design
}
\examples{
\dontrun{
ggplot(...) +
ssz_theme()
}
}
