% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssz_theme_void.R
\name{ssz_theme_void}
\alias{ssz_theme_void}
\title{ssz_theme_void}
\usage{
ssz_theme_void(
  base_size = 8.2,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_size}{optional, basic font size, parameter passed to minimal theme}

\item{base_family}{optional, basic font family, parameter to be passed to minimal theme}

\item{base_line_size}{optional, basic line size, parameter passed to minimal theme}

\item{base_rect_size}{optional, basic rect size, parameter passed to minimal theme}
}
\description{
Function for styling void ggplots according to Stadt Zürich's corporate design
}
\examples{
\dontrun{
ggplot(...) +
  ssz_theme_void()
}
}
